+++ src/Sylius/Bundle/ShippingBundle/spec/Form/Type/Calculator/FlatRateConfigurationTypeSpec.php
@@ -55,6 +55,6 @@ class FlatRateConfigurationTypeSpec extends ObjectBehavior
 
     function it_has_a_name()
     {
-        $this->getName()->shouldReturn('sylius_shipping_calculator_flat_rate_configuration');
+        $this->getName()->shouldReturn('sylius_shipping_calculator_flat_rate');
     }
 }