+++ src/Sylius/Bundle/ShippingBundle/spec/Form/Type/Calculator/FlexibleRateConfigurationTypeSpec.php
@@ -57,6 +57,6 @@ class FlexibleRateConfigurationTypeSpec extends ObjectBehavior
 
     function it_has_a_name()
     {
-        $this->getName()->shouldReturn('sylius_shipping_calculator_flexible_rate_configuration');
+        $this->getName()->shouldReturn('sylius_shipping_calculator_flexible_rate');
     }
 }