+++ src/Sylius/Bundle/ShippingBundle/spec/Form/Type/Calculator/VolumeRateConfigurationTypeSpec.php
@@ -56,6 +56,6 @@ class VolumeRateConfigurationTypeSpec extends ObjectBehavior
 
     function it_has_a_name()
     {
-        $this->getName()->shouldReturn('sylius_shipping_calculator_volume_rate_configuration');
+        $this->getName()->shouldReturn('sylius_shipping_calculator_volume_rate');
     }
 }